requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
anthropic>=0.39.0
python-multipart>=0.0.6
bcrypt>=4.0.0
PyJWT>=2.8.0
stripe>=7.0.0
